modified

+++ HASHING/HASHING.java
@@ -14,10 +14,10 @@ public class HASHING {
         for (int i = 0; i < n; i++)
             res += arr[i];
         System.out.println(res);
-        double startOne = (0);
-        double endOne = ((n / 2f) - 1);
-        double startTwo = ((n / 2f));
-        double endTwo = ((n - 1));
+        int  startOne = (0);
+        int  endOne = (int)((n / 2f) - 1);
+        int  startTwo =(int) ((n / 2f));
+        int  endTwo = ((n - 1));
         System.out.println(startOne);
         System.out.println(endOne);
         System.out.println(startTwo);
